requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
requests>=2.31.0
atlassian-python-api>=3.41.0

